refactor(agentic-ai): move new models to a domain package
