+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/domain/model/message/ContentMessage.java
@@ -4,9 +4,9 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.aiagent.model.message;
+package io.camunda.connector.agenticai.domain.model.message;
 
-import io.camunda.connector.agenticai.aiagent.model.message.content.Content;
+import io.camunda.connector.agenticai.domain.model.message.content.Content;
 import java.util.List;
 
 public interface ContentMessage {
