+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/domain/model/message/MessageRole.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.aiagent.model.message;
+package io.camunda.connector.agenticai.domain.model.message;
 
 import com.fasterxml.jackson.annotation.JsonValue;
