+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/domain/model/message/content/TextContent.java
@@ -4,20 +4,19 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.aiagent.model.message.content;
+package io.camunda.connector.agenticai.domain.model.message.content;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import io.camunda.document.Document;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record DocumentContent(Document document) implements Content {
-  public DocumentContent {
-    if (document == null) {
-      throw new IllegalArgumentException("Document cannot be null");
+public record TextContent(String text) implements Content {
+  public TextContent {
+    if (text == null || text.isBlank()) {
+      throw new IllegalArgumentException("Text cannot be null or empty");
     }
   }
 
-  public static DocumentContent documentContent(Document document) {
-    return new DocumentContent(document);
+  public static TextContent textContent(String text) {
+    return new TextContent(text);
   }
 }
