+++ connectors/agentic-ai/src/main/java/io/camunda/connector/agenticai/domain/model/tools/ToolCallResult.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.agenticai.aiagent.model.message.tools;
+package io.camunda.connector.agenticai.domain.model.tools;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
